func to generate id unique TxnRef

+++ src/main/java/com/capstone_project/hbts/dto/Payment/TransactionDTO.java
@@ -25,6 +25,6 @@ public class TransactionDTO {
 
     private String transactionNo;
 
-    private int idService; // unique
+    private long idService; // unique
 
 }
